requirements.txt
bleach==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
kaggle==1.7.4.5
protobuf==6.32.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
requests==2.32.5
six==1.17.0
text-unidecode==1.3
tqdm==4.67.1
urllib3==2.5.0
webencodings==0.5.1
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.1
importlib_resources==6.5.2
joblib==1.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyparsing==3.2.5
pytz==2025.2
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
threadpoolctl==3.6.0
tzdata==2025.2
zipp==3.23.0
imageio==2.37.0
lazy-loader==0.4
networkx==3.2.1
scikit-image==0.24.0
tifffile==2024.8.30
